move logic to findWordBoundries
